pyproject.toml
[tool.poetry]
name = "port-scanner"
version = "0.1.0"
license = "MIT"
description = "Python module for scaning ports using nmap"
authors = ["Антон Григорьев <grianton535@gmail.com>"]
repository = "https://github.com/zeph1rrinc/port_scanner"
readme = "README.md"
packages = [{include = "port_scanner"}]

[tool.poetry.dependencies]
python = "^3.10"
python-nmap = "^0.7.1"
loguru = "^0.6.0"

[tool.poetry.scripts]
start = "port_scanner.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
